+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixed
+
+- `DiscourseTheme::Watcher` was incorrectly ignoring any files with `migrations` in the pathname. (#39)
+- `DiscourseTheme::Watcher` was incorrectly uploading migration files when any file changes have been detected. (#39)
+
 ## [1.1.0] - 2024-01-10
 
 ### Added